fixed multiple GCC warnings on Ubuntu 11.04

+++ 3rdparty/libjasper/jas_icc.c
@@ -730,9 +730,9 @@ void jas_iccattrtab_dump(jas_iccattrtab_t *attrtab, FILE *out)
 		fprintf(out, "attrno=%d; attrname=\"%s\"(0x%08x); attrtype=\"%s\"(0x%08x)\n",
 		  i,
 		  jas_iccsigtostr(attr->name, &buf[0]),
-		  attr->name,
+		  (unsigned)attr->name,
 		  jas_iccsigtostr(attrval->type, &buf[8]),
-		  attrval->type
+		  (unsigned)attrval->type
 		  );
 		jas_iccattrval_dump(attrval, out);
 		fprintf(out, "---\n");
@@ -884,7 +884,7 @@ void jas_iccattrval_dump(jas_iccattrval_t *attrval, FILE *out)
 	char buf[8];
 	jas_iccsigtostr(attrval->type, buf);
 	fprintf(out, "refcnt = %d; type = 0x%08x %s\n", attrval->refcnt,
-	  attrval->type, jas_iccsigtostr(attrval->type, &buf[0]));
+	  (unsigned)attrval->type, jas_iccsigtostr(attrval->type, &buf[0]));
 	if (attrval->ops->dump) {
 		(*attrval->ops->dump)(attrval, out);
 	}
@@ -1041,7 +1041,7 @@ static void jas_icccurv_dump(jas_iccattrval_t *attrval, FILE *out)
 {
 	int i;
 	jas_icccurv_t *curv = &attrval->data.curv;
-	fprintf(out, "number of entires = %d\n", curv->numents);
+	fprintf(out, "number of entires = %d\n", (int)curv->numents);
 	if (curv->numents == 1) {
 		fprintf(out, "gamma = %f\n", curv->ents[0] / 256.0);
 	} else {
@@ -1167,9 +1167,9 @@ static void jas_icctxtdesc_dump(jas_iccattrval_t *attrval, FILE *out)
 {
 	jas_icctxtdesc_t *txtdesc = &attrval->data.txtdesc;
 	fprintf(out, "ascii = \"%s\"\n", txtdesc->ascdata);
-	fprintf(out, "uclangcode = %d; uclen = %d\n", txtdesc->uclangcode,
-	  txtdesc->uclen);
-	fprintf(out, "sccode = %d\n", txtdesc->sccode);
+	fprintf(out, "uclangcode = %d; uclen = %d\n", (int)txtdesc->uclangcode,
+	  (int)txtdesc->uclen);
+	fprintf(out, "sccode = %d\n", (int)txtdesc->sccode);
 	fprintf(out, "maclen = %d\n", txtdesc->maclen);
 }
 
@@ -1401,7 +1401,7 @@ static void jas_icclut8_dump(jas_iccattrval_t *attrval, FILE *out)
 		fprintf(out, "\n");
 	}
 	fprintf(out, "numintabents=%d, numouttabents=%d\n",
-	  lut8->numintabents, lut8->numouttabents);
+	  (int)lut8->numintabents, (int)lut8->numouttabents);
 }
 
 /******************************************************************************\
@@ -1565,7 +1565,7 @@ static void jas_icclut16_dump(jas_iccattrval_t *attrval, FILE *out)
 		fprintf(out, "\n");
 	}
 	fprintf(out, "numintabents=%d, numouttabents=%d\n",
-	  lut16->numintabents, lut16->numouttabents);
+	  (int)lut16->numintabents, (int)lut16->numouttabents);
 }
 
 /******************************************************************************\
