+++ 3rdparty/libjasper/jp2_cod.c
@@ -321,7 +321,7 @@ void jp2_box_dump(jp2_box_t *box, FILE *out)
 
 	fprintf(out, "JP2 box: ");
 	fprintf(out, "type=%c%s%c (0x%08x); length=%d\n", '"', boxinfo->name,
-	  '"', box->type, box->len);
+	  '"', (unsigned)box->type, (int)box->len);
 	if (box->ops->dumpdata) {
 		(*box->ops->dumpdata)(box, out);
 	}
@@ -433,7 +433,7 @@ static void jp2_cdef_dumpdata(jp2_box_t *box, FILE *out)
 	unsigned int i;
 	for (i = 0; i < cdef->numchans; ++i) {
 		fprintf(out, "channo=%d; type=%d; assoc=%d\n",
-		  cdef->ents[i].channo, cdef->ents[i].type, cdef->ents[i].assoc);
+		  (int)cdef->ents[i].channo, (int)cdef->ents[i].type, (int)cdef->ents[i].assoc);
 	}
 }
 
@@ -871,7 +871,7 @@ static void jp2_pclr_dumpdata(jp2_box_t *box, FILE *out)
 	  (int) pclr->numchans);
 	for (i = 0; i < pclr->numlutents; ++i) {
 		for (j = 0; j < pclr->numchans; ++j) {
-			fprintf(out, "LUT[%d][%d]=%d\n", i, j, pclr->lutdata[i * pclr->numchans + j]);
+			fprintf(out, "LUT[%d][%d]=%d\n", i, j, (int)pclr->lutdata[i * pclr->numchans + j]);
 		}
 	}
 }
